requirements.txt
aiofiles==23.2.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
black==24.2.0
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
docstring-parser==0.15
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.4
httpx==0.27.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.6
inflection==0.5.1
Jinja2==3.1.3
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
piccolo==1.3.2
piccolo-api==1.1.0
piccolo_admin==1.2.2
platformdirs==4.2.0
priority==2.0.0
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
python-decouple==3.8
python-multipart==0.0.9
sniffio==1.3.1
starlette==0.36.3
targ==0.3.8
taskgroup==0.0.0a4
tomli==2.0.1
typing_extensions==4.10.0
uvicorn==0.27.1
wsproto==1.2.0
